+++ packages/solana-wallet-snap/src/core/services/signer/mocks/scenarios/types.ts
@@ -1,17 +1,17 @@
+import type { ExtendedKeyringAccount } from '@metamask/snap-networks-utils';
 import type {
   TransactionMessage,
   TransactionMessageWithBlockhashLifetime,
   TransactionMessageWithFeePayer,
 } from '@solana/kit';
 
-import type { SolanaKeyringAccount } from '../../../../../entities';
 import type { Network } from '../../../../constants/solana';
 
 export type MockExecutionScenario = {
   name: string;
   scope: Network;
-  fromAccount: SolanaKeyringAccount;
-  toAccount: SolanaKeyringAccount;
+  fromAccount: ExtendedKeyringAccount;
+  toAccount: ExtendedKeyringAccount;
   fromAccountPrivateKeyBytes: Uint8Array;
   transactionMessage: TransactionMessage &
     TransactionMessageWithFeePayer &